requirements.txt
Django>=5.0.0,<6.0
djangorestframework>=3.12.0,<4.0
drf-spectacular>=0.27.0,<0.28
django-jazzmin>=2.6.0,<3.0
